+++ src/pages/IntroPage.tsx
@@ -1,20 +1,10 @@
-import { useNavigate } from 'react-router';
-
-import Button from '../components/common/Button';
+import OrderButtons from '../components/OrderButtons';
 
 export default function IntroPage() {
-  const navigation = useNavigate();
-  const HandelOnClick = () => {
-    navigation('/order');
-  };
-
   return (
     <div>
       <h2>원하시는 주문을 터치해주세요</h2>
-      <div>
-        <Button text="🙋🏻‍♀️ 매장 주문" onClick={HandelOnClick} />
-        <Button text="🛍️ 전체 포장" onClick={HandelOnClick} />
-      </div>
+      <OrderButtons />
     </div>
   );
 }
